More updates for 64 bit machines - replace "long"s with "int"s     unless the long is used in pointer arithmetic. Add DSMSensor::get/setTypeName in order to support documenting a     long-winded sensor type name in the config. Fix some broken tests.

+++ src/nidas/core/SampleIOProcessor.cc
@@ -230,7 +230,7 @@ void SampleIOProcessor::fromDOMElement(const xercesc::DOMElement* node)
                // If you unset the dec flag, then a leading '0' means
                // octal, and 0x means hex.
                ist.unsetf(ios::dec);
-               unsigned long val;
+               unsigned int val;
                ist >> val;
                if (ist.fail())
                    throw n_u::InvalidParameterException("sensor",
